requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
pandas==1.5.1
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
